requirements.txt
pyTelegramBotAPI==4.25.0
psycopg2
python-dotenv
numpy
